Updated ToDo list. Added more info to MS/TP capture utility readme file. Changed silence timer in BDK to compile smaller.

+++ trunk/bacnet-stack/ports/bdk-atxx4-mstp/rs485.h
@@ -63,7 +63,7 @@ extern "C" {
     void rs485_silence_time_reset(
         void);
     bool rs485_silence_time_elapsed(
-        uint32_t milliseconds);
+        uint16_t milliseconds);
 
 #ifdef __cplusplus
 }
